Allow Godot to make external requests
We must not add the "https://{CLIENT_ID}.discordsays.com/.proxy/" prefix if
the url already includes it. Since Godot does not support relative urls, this
would otherwise make it impossible to use HttpClient / HttpRequest.

Additionally, if the server uses gzip encoding, either the Discord sandbox or
this fetch override causes the body to already be decompressed, but the
Content-Encoding header to still be in place. This again causes the request to
fail in Godot because it will attempt to decompress it again. This change also
removes any Content-Encoding headers from the response for that reason.